編集削除にオーナー判定

+++ mern-stack-backend/index.js
@@ -48,28 +48,28 @@ app.get('/item/:id', async (req, res) => {
 
 // Update Item
 app.put('/item/update/:id', auth, async (req, res) => {
-  let message = 'アイテム編集成功';
   try {
     await connectDB();
-    const result = await ItemModel.updateOne({ _id: req.params.id }, req.body);
-    return res.status(200).json({ message, result });
-  } catch (error) {
-    message = 'アイテム編集失敗';
-    return res.status(400).json({ message });
-  }
+    const singleItem = await ItemModel.findById(req.params.id);
+    if (singleItem.email === req.body.email) {
+      await ItemModel.updateOne({ _id: req.params.id }, req.body);
+      return res.status(200).json({ message: 'アイテム編集成功' });
+    }
+  } catch (error) {}
+  return res.status(400).json({ message: 'アイテム編集失敗' });
 });
 
 // Delete Item
 app.delete('/item/delete/:id', auth, async (req, res) => {
-  let message = 'アイテム削除成功';
   try {
     await connectDB();
-    const result = await ItemModel.deleteOne({ _id: req.params.id });
-    return res.status(200).json({ message, result });
-  } catch (error) {
-    message = 'アイテム削除失敗';
-    return res.status(400).json({ message });
-  }
+    const singleItem = await ItemModel.findById(req.params.id);
+    if (singleItem.email === req.body.email) {
+      await ItemModel.deleteOne({ _id: req.params.id });
+      return res.status(200).json({ message: 'アイテム削除成功' });
+    }
+  } catch (error) {}
+  return res.status(400).json({ message: 'アイテム削除失敗' });
 });
 
 // USER function
